Extract AppLanguage model from settings store
