+++ app/src/main/swift/Models/SettingsStore.swift
@@ -63,1423 +63,6 @@ enum VirtualPadSkin: Int, CaseIterable, Identifiable {
     }
 }
 
-enum AppLanguage: String, CaseIterable, Identifiable {
-    case system
-    case english
-    case simplifiedChinese
-    case arabic
-    case spanish
-    case french
-    case portuguese
-    case japanese
-    case korean
-
-    var id: String { rawValue }
-
-    var label: String {
-        switch self {
-        case .system: return "System Default"
-        case .english: return "English"
-        case .simplifiedChinese: return "简体中文"
-        case .arabic: return "العربية"
-        case .spanish: return "Español"
-        case .french: return "Français"
-        case .portuguese: return "Português"
-        case .japanese: return "日本語"
-        case .korean: return "한국어"
-        }
-    }
-
-    static func resolvedSystemLanguage() -> AppLanguage {
-        let code = Locale.current.language.languageCode?.identifier.lowercased() ?? "en"
-        switch code {
-        case "zh":
-            return .simplifiedChinese
-        case "ar":
-            return .arabic
-        case "es":
-            return .spanish
-        case "fr":
-            return .french
-        case "pt":
-            return .portuguese
-        case "ja":
-            return .japanese
-        case "ko":
-            return .korean
-        default:
-            return .english
-        }
-    }
-
-    var resolved: AppLanguage {
-        self == .system ? Self.resolvedSystemLanguage() : self
-    }
-
-    var layoutDirection: LayoutDirection {
-        resolved == .arabic ? .rightToLeft : .leftToRight
-    }
-
-    func localized(_ key: String) -> String {
-        guard let translated = Self.translations[resolved]?[key] else {
-            return Self.commonTranslations[resolved]?[key] ?? Self.uiSupplementTranslations[resolved]?[key] ?? key
-        }
-        return translated
-    }
-
-    private static let translations: [AppLanguage: [String: String]] = [
-        .simplifiedChinese: [
-            "System Default": "跟随系统",
-            "Language": "语言",
-            "Interface Language": "界面语言",
-            "App Language": "应用语言",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "ARMSX2 iOS 菜单会尽量使用此语言。部分模拟器术语和调试消息可能仍以英文显示。",
-            "Games": "游戏",
-            "BIOS": "BIOS",
-            "Help": "帮助",
-            "Settings": "设置",
-            "File Import": "文件导入",
-            "Import Result": "导入结果",
-            "Cover Result": "封面结果",
-            "Cover Source": "封面来源",
-            "OK": "确定",
-            "Cancel": "取消",
-            "Save": "保存",
-            "Done": "完成",
-            "Restart": "重启",
-            "Stop": "停止",
-            "Resume": "继续",
-            "Version": "版本",
-            "About": "关于",
-            "Import Games": "导入游戏",
-            "Import BIOS": "导入 BIOS",
-            "BIOS Only": "仅启动 BIOS",
-            "No Games Found": "未找到游戏",
-            "Import PS2 disc images to add them here.": "导入 PS2 光盘镜像后会显示在这里。",
-            "No BIOS Found": "未找到 BIOS",
-            "Import a PS2 BIOS dump to enable booting.": "导入 PS2 BIOS 转储文件以启用启动。",
-            "Compatibility Picker": "兼容选择器",
-            "If one picker refuses to select your .bin/.rom file, try the other.": "如果一个选择器无法选择 .bin/.rom 文件，请尝试另一个。",
-            "Emulator": "模拟器",
-            "Graphics": "图形",
-            "Network": "网络",
-            "Memory Cards": "记忆卡",
-            "Storage": "存储",
-            "RetroAchievements": "RetroAchievements",
-            "Overlay (OSD)": "叠加层 (OSD)",
-            "Game Controller": "游戏手柄",
-            "Local Multiplayer": "本地多人",
-            "Virtual Pad": "虚拟按键",
-            "Licenses & Credits": "许可证与鸣谢",
-            "Performance Overlay": "性能叠加层",
-            "Preset": "预设",
-            "Position": "位置",
-            "Hidden": "隐藏",
-            "Top Left": "左上",
-            "Top Right": "右上",
-            "Displayed Items": "显示项目",
-            "Show FPS": "显示 FPS",
-            "Show VPS": "显示 VPS",
-            "Show Speed": "显示速度",
-            "Show CPU": "显示 CPU",
-            "Show GPU": "显示 GPU",
-            "Show Resolution": "显示分辨率",
-            "Show GS Stats": "显示 GS 统计",
-            "Show Indicators": "显示指示器",
-            "Show Settings": "显示设置",
-            "Show Inputs": "显示输入",
-            "Show Frame Times": "显示帧时间",
-            "Show Version": "显示版本",
-            "Show Hardware Info": "显示硬件信息",
-            "Show Device Stats": "显示设备状态",
-            "Notes": "说明",
-            "Overlay": "叠加层",
-            "Appearance": "外观",
-            "Button Skin": "按键皮肤",
-            "Custom Skin": "自定义皮肤",
-            "Import Button Images": "导入按键图片",
-            "Feedback": "反馈",
-            "Haptic Feedback": "触觉反馈",
-            "Layout": "布局",
-            "Edit Layout": "编辑布局",
-            "Drag buttons to reposition. Pinch to resize.": "拖动按键调整位置，捏合调整大小。",
-            "Custom Imported": "自定义导入",
-            "Crisp Vector": "清晰矢量",
-            "ARMSX2 Refresh": "ARMSX2 Refresh",
-            "Settings Guide": "设置指南",
-            "Supported Formats": "支持格式"
-        ],
-        .arabic: [
-            "System Default": "حسب النظام",
-            "Language": "اللغة",
-            "Interface Language": "لغة الواجهة",
-            "App Language": "لغة التطبيق",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "ستستخدم قوائم ARMSX2 iOS هذه اللغة عند توفرها. قد تظل بعض مصطلحات المحاكي ورسائل التصحيح باللغة الإنجليزية.",
-            "Games": "الألعاب",
-            "BIOS": "BIOS",
-            "Help": "مساعدة",
-            "Settings": "الإعدادات",
-            "File Import": "استيراد ملف",
-            "Import Result": "نتيجة الاستيراد",
-            "Cover Result": "نتيجة الغلاف",
-            "Cover Source": "مصدر الأغلفة",
-            "OK": "موافق",
-            "Cancel": "إلغاء",
-            "Save": "حفظ",
-            "Done": "تم",
-            "Restart": "إعادة تشغيل",
-            "Stop": "إيقاف",
-            "Resume": "استئناف",
-            "Version": "الإصدار",
-            "About": "حول",
-            "Import Games": "استيراد الألعاب",
-            "Import BIOS": "استيراد BIOS",
-            "BIOS Only": "BIOS فقط",
-            "No Games Found": "لم يتم العثور على ألعاب",
-            "Import PS2 disc images to add them here.": "استورد صور أقراص PS2 لإضافتها هنا.",
-            "No BIOS Found": "لم يتم العثور على BIOS",
-            "Import a PS2 BIOS dump to enable booting.": "استورد ملف BIOS من جهاز PS2 لتفعيل التشغيل.",
-            "Compatibility Picker": "منتقي التوافق",
-            "If one picker refuses to select your .bin/.rom file, try the other.": "إذا رفض أحد المنتقيات اختيار ملف .bin/.rom، جرّب الآخر.",
-            "Emulator": "المحاكي",
-            "Graphics": "الرسوميات",
-            "Network": "الشبكة",
-            "Memory Cards": "بطاقات الذاكرة",
-            "Storage": "التخزين",
-            "RetroAchievements": "RetroAchievements",
-            "Overlay (OSD)": "التراكب (OSD)",
-            "Game Controller": "يد التحكم",
-            "Local Multiplayer": "لعب محلي متعدد",
-            "Virtual Pad": "لوحة افتراضية",
-            "Licenses & Credits": "التراخيص والشكر",
-            "Performance Overlay": "تراكب الأداء",
-            "Preset": "الإعداد المسبق",
-            "Position": "الموضع",
-            "Hidden": "مخفي",
-            "Top Left": "أعلى اليسار",
-            "Top Right": "أعلى اليمين",
-            "Displayed Items": "العناصر المعروضة",
-            "Show FPS": "إظهار FPS",
-            "Show VPS": "إظهار VPS",
-            "Show Speed": "إظهار السرعة",
-            "Show CPU": "إظهار CPU",
-            "Show GPU": "إظهار GPU",
-            "Show Resolution": "إظهار الدقة",
-            "Show GS Stats": "إظهار إحصاءات GS",
-            "Show Indicators": "إظهار المؤشرات",
-            "Show Settings": "إظهار الإعدادات",
-            "Show Inputs": "إظهار الإدخال",
-            "Show Frame Times": "إظهار أزمنة الإطارات",
-            "Show Version": "إظهار الإصدار",
-            "Show Hardware Info": "إظهار معلومات العتاد",
-            "Show Device Stats": "إظهار حالة الجهاز",
-            "Notes": "ملاحظات",
-            "Overlay": "التراكب",
-            "Appearance": "المظهر",
-            "Button Skin": "مظهر الأزرار",
-            "Custom Skin": "مظهر مخصص",
-            "Import Button Images": "استيراد صور الأزرار",
-            "Feedback": "التغذية الراجعة",
-            "Haptic Feedback": "اهتزاز لمسي",
-            "Layout": "التخطيط",
-            "Edit Layout": "تعديل التخطيط",
-            "Drag buttons to reposition. Pinch to resize.": "اسحب الأزرار لتغيير موضعها، واقرص لتغيير الحجم.",
-            "Custom Imported": "مخصص مستورد",
-            "Crisp Vector": "متجه حاد",
-            "ARMSX2 Refresh": "ARMSX2 Refresh",
-            "Settings Guide": "دليل الإعدادات",
-            "Supported Formats": "الصيغ المدعومة"
-        ],
-        .spanish: [
-            "System Default": "Predeterminado del sistema",
-            "Language": "Idioma",
-            "Interface Language": "Idioma de la interfaz",
-            "App Language": "Idioma de la app",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "Los menús de ARMSX2 iOS usarán este idioma cuando esté disponible. Algunos términos del emulador y mensajes de depuración pueden seguir apareciendo en inglés.",
-            "Games": "Juegos",
-            "BIOS": "BIOS",
-            "Help": "Ayuda",
-            "Settings": "Ajustes",
-            "File Import": "Importar archivo",
-            "Import Result": "Resultado de importación",
-            "Cover Result": "Resultado de carátula",
-            "Cover Source": "Fuente de carátulas",
-            "OK": "Aceptar",
-            "Cancel": "Cancelar",
-            "Save": "Guardar",
-            "Done": "Listo",
-            "Restart": "Reiniciar",
-            "Stop": "Detener",
-            "Resume": "Continuar",
-            "Version": "Versión",
-            "About": "Acerca de",
-            "Import Games": "Importar juegos",
-            "Import BIOS": "Importar BIOS",
-            "BIOS Only": "Solo BIOS",
-            "No Games Found": "No se encontraron juegos",
-            "Import PS2 disc images to add them here.": "Importa imágenes de disco de PS2 para añadirlas aquí.",
-            "No BIOS Found": "No se encontró BIOS",
-            "Import a PS2 BIOS dump to enable booting.": "Importa una copia de BIOS de PS2 para poder iniciar.",
-            "Compatibility Picker": "Selector de compatibilidad",
-            "If one picker refuses to select your .bin/.rom file, try the other.": "Si un selector no permite elegir tu .bin/.rom, prueba el otro.",
-            "Emulator": "Emulador",
-            "Graphics": "Gráficos",
-            "Network": "Red",
-            "Memory Cards": "Tarjetas de memoria",
-            "Storage": "Almacenamiento",
-            "RetroAchievements": "RetroAchievements",
-            "Overlay (OSD)": "Superposición (OSD)",
-            "Game Controller": "Mando",
-            "Local Multiplayer": "Multijugador local",
-            "Virtual Pad": "Controles táctiles",
-            "Licenses & Credits": "Licencias y créditos",
-            "Performance Overlay": "Superposición de rendimiento",
-            "Preset": "Preset",
-            "Position": "Posición",
-            "Hidden": "Oculto",
-            "Top Left": "Arriba izquierda",
-            "Top Right": "Arriba derecha",
-            "Displayed Items": "Elementos mostrados",
-            "Show FPS": "Mostrar FPS",
-            "Show VPS": "Mostrar VPS",
-            "Show Speed": "Mostrar velocidad",
-            "Show CPU": "Mostrar CPU",
-            "Show GPU": "Mostrar GPU",
-            "Show Resolution": "Mostrar resolución",
-            "Show GS Stats": "Mostrar estadísticas GS",
-            "Show Indicators": "Mostrar indicadores",
-            "Show Settings": "Mostrar ajustes",
-            "Show Inputs": "Mostrar entradas",
-            "Show Frame Times": "Mostrar tiempos de fotograma",
-            "Show Version": "Mostrar versión",
-            "Show Hardware Info": "Mostrar hardware",
-            "Show Device Stats": "Mostrar estado del dispositivo",
-            "Notes": "Notas",
-            "Overlay": "Superposición",
-            "Appearance": "Apariencia",
-            "Button Skin": "Tema de botones",
-            "Custom Skin": "Tema personalizado",
-            "Import Button Images": "Importar imágenes de botones",
-            "Feedback": "Respuesta",
-            "Haptic Feedback": "Respuesta háptica",
-            "Layout": "Diseño",
-            "Edit Layout": "Editar diseño",
-            "Drag buttons to reposition. Pinch to resize.": "Arrastra los botones para moverlos. Pellizca para cambiar tamaño.",
-            "Custom Imported": "Personalizado importado",
-            "Crisp Vector": "Vector nítido",
-            "ARMSX2 Refresh": "ARMSX2 Refresh",
-            "Settings Guide": "Guía de ajustes",
-            "Supported Formats": "Formatos compatibles"
-        ],
-        .french: [
-            "System Default": "Langue du système",
-            "Language": "Langue",
-            "Interface Language": "Langue de l'interface",
-            "App Language": "Langue de l'app",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "Les menus d'ARMSX2 iOS utiliseront cette langue si disponible. Certains termes de l'émulateur et messages de débogage peuvent rester en anglais.",
-            "Games": "Jeux",
-            "BIOS": "BIOS",
-            "Help": "Aide",
-            "Settings": "Réglages",
-            "OK": "OK",
-            "Cancel": "Annuler",
-            "Save": "Enregistrer",
-            "Done": "Terminé",
-            "Restart": "Redémarrer",
-            "Stop": "Arrêter",
-            "Resume": "Reprendre",
-            "Version": "Version",
-            "About": "À propos",
-            "Import Games": "Importer des jeux",
-            "Import BIOS": "Importer le BIOS",
-            "No Games Found": "Aucun jeu trouvé",
-            "No BIOS Found": "Aucun BIOS trouvé",
-            "Emulator": "Émulateur",
-            "Graphics": "Graphismes",
-            "Network": "Réseau",
-            "Memory Cards": "Cartes mémoire",
-            "Storage": "Stockage",
-            "Overlay (OSD)": "Superposition (OSD)",
-            "Game Controller": "Manette",
-            "Local Multiplayer": "Multijoueur local",
-            "Virtual Pad": "Commandes tactiles",
-            "Licenses & Credits": "Licences et crédits",
-            "Performance Overlay": "Superposition de performance",
-            "Preset": "Préréglage",
-            "Position": "Position",
-            "Hidden": "Masqué",
-            "Top Left": "En haut à gauche",
-            "Top Right": "En haut à droite",
-            "Displayed Items": "Éléments affichés",
-            "Show FPS": "Afficher FPS",
-            "Show CPU": "Afficher CPU",
-            "Show GPU": "Afficher GPU",
-            "Show Device Stats": "Afficher l'état de l'appareil",
-            "Overlay": "Superposition",
-            "Appearance": "Apparence",
-            "Button Skin": "Thème des boutons",
-            "Custom Skin": "Thème personnalisé",
-            "Import Button Images": "Importer des images de boutons",
-            "Haptic Feedback": "Retour haptique",
-            "Edit Layout": "Modifier la disposition"
-        ],
-        .portuguese: [
-            "System Default": "Padrão do sistema",
-            "Language": "Idioma",
-            "Interface Language": "Idioma da interface",
-            "App Language": "Idioma do app",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "Os menus do ARMSX2 iOS usarão este idioma quando disponível. Alguns termos do emulador e mensagens de depuração ainda podem aparecer em inglês.",
-            "Games": "Jogos",
-            "BIOS": "BIOS",
-            "Help": "Ajuda",
-            "Settings": "Configurações",
-            "OK": "OK",
-            "Cancel": "Cancelar",
-            "Save": "Salvar",
-            "Done": "Concluído",
-            "Restart": "Reiniciar",
-            "Stop": "Parar",
-            "Resume": "Continuar",
-            "Version": "Versão",
-            "About": "Sobre",
-            "Import Games": "Importar jogos",
-            "Import BIOS": "Importar BIOS",
-            "No Games Found": "Nenhum jogo encontrado",
-            "No BIOS Found": "Nenhuma BIOS encontrada",
-            "Emulator": "Emulador",
-            "Graphics": "Gráficos",
-            "Network": "Rede",
-            "Memory Cards": "Cartões de memória",
-            "Storage": "Armazenamento",
-            "Overlay (OSD)": "Sobreposição (OSD)",
-            "Game Controller": "Controle",
-            "Local Multiplayer": "Multijogador local",
-            "Virtual Pad": "Controles virtuais",
-            "Licenses & Credits": "Licenças e créditos",
-            "Performance Overlay": "Sobreposição de desempenho",
-            "Preset": "Predefinição",
-            "Position": "Posição",
-            "Hidden": "Oculto",
-            "Top Left": "Superior esquerdo",
-            "Top Right": "Superior direito",
-            "Displayed Items": "Itens exibidos",
-            "Show FPS": "Mostrar FPS",
-            "Show CPU": "Mostrar CPU",
-            "Show GPU": "Mostrar GPU",
-            "Show Device Stats": "Mostrar status do dispositivo",
-            "Overlay": "Sobreposição",
-            "Appearance": "Aparência",
-            "Button Skin": "Tema dos botões",
-            "Custom Skin": "Tema personalizado",
-            "Import Button Images": "Importar imagens dos botões",
-            "Haptic Feedback": "Feedback tátil",
-            "Edit Layout": "Editar layout"
-        ],
-        .japanese: [
-            "System Default": "システム設定",
-            "Language": "言語",
-            "Interface Language": "インターフェース言語",
-            "App Language": "アプリの言語",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "ARMSX2 iOS のメニューは、利用可能な場合この言語を使用します。一部のエミュレーター用語やデバッグメッセージは英語のまま表示される場合があります。",
-            "Games": "ゲーム",
-            "BIOS": "BIOS",
-            "Help": "ヘルプ",
-            "Settings": "設定",
-            "OK": "OK",
-            "Cancel": "キャンセル",
-            "Save": "保存",
-            "Done": "完了",
-            "Restart": "再起動",
-            "Stop": "停止",
-            "Resume": "再開",
-            "Version": "バージョン",
-            "About": "情報",
-            "Import Games": "ゲームをインポート",
-            "Import BIOS": "BIOSをインポート",
-            "No Games Found": "ゲームが見つかりません",
-            "No BIOS Found": "BIOSが見つかりません",
-            "Emulator": "エミュレーター",
-            "Graphics": "グラフィック",
-            "Network": "ネットワーク",
-            "Memory Cards": "メモリーカード",
-            "Storage": "ストレージ",
-            "Overlay (OSD)": "オーバーレイ (OSD)",
-            "Game Controller": "ゲームコントローラー",
-            "Local Multiplayer": "ローカルマルチプレイ",
-            "Virtual Pad": "仮想パッド",
-            "Licenses & Credits": "ライセンスとクレジット",
-            "Performance Overlay": "パフォーマンス表示",
-            "Preset": "プリセット",
-            "Position": "位置",
-            "Hidden": "非表示",
-            "Top Left": "左上",
-            "Top Right": "右上",
-            "Displayed Items": "表示項目",
-            "Show FPS": "FPSを表示",
-            "Show CPU": "CPUを表示",
-            "Show GPU": "GPUを表示",
-            "Show Device Stats": "デバイス状態を表示",
-            "Overlay": "オーバーレイ",
-            "Appearance": "外観",
-            "Button Skin": "ボタンスキン",
-            "Custom Skin": "カスタムスキン",
-            "Import Button Images": "ボタン画像をインポート",
-            "Haptic Feedback": "触覚フィードバック",
-            "Edit Layout": "レイアウト編集"
-        ],
-        .korean: [
-            "System Default": "시스템 기본값",
-            "Language": "언어",
-            "Interface Language": "인터페이스 언어",
-            "App Language": "앱 언어",
-            "ARMSX2 iOS menus will use this language where available. Some emulator terms and debug messages may still appear in English.": "ARMSX2 iOS 메뉴는 가능한 경우 이 언어를 사용합니다. 일부 에뮬레이터 용어와 디버그 메시지는 영어로 표시될 수 있습니다.",
-            "Games": "게임",
-            "BIOS": "BIOS",
-            "Help": "도움말",
-            "Settings": "설정",
-            "OK": "확인",
-            "Cancel": "취소",
-            "Save": "저장",
-            "Done": "완료",
-            "Restart": "재시작",
-            "Stop": "중지",
-            "Resume": "재개",
-            "Version": "버전",
-            "About": "정보",
-            "Import Games": "게임 가져오기",
-            "Import BIOS": "BIOS 가져오기",
-            "No Games Found": "게임을 찾을 수 없음",
-            "No BIOS Found": "BIOS를 찾을 수 없음",
-            "Emulator": "에뮬레이터",
-            "Graphics": "그래픽",
-            "Network": "네트워크",
-            "Memory Cards": "메모리 카드",
-            "Storage": "저장 공간",
-            "Overlay (OSD)": "오버레이 (OSD)",
-            "Game Controller": "게임 컨트롤러",
-            "Local Multiplayer": "로컬 멀티플레이",
-            "Virtual Pad": "가상 패드",
-            "Licenses & Credits": "라이선스 및 크레딧",
-            "Performance Overlay": "성능 오버레이",
-            "Preset": "프리셋",
-            "Position": "위치",
-            "Hidden": "숨김",
-            "Top Left": "왼쪽 위",
-            "Top Right": "오른쪽 위",
-            "Displayed Items": "표시 항목",
-            "Show FPS": "FPS 표시",
-            "Show CPU": "CPU 표시",
-            "Show GPU": "GPU 표시",
-            "Show Device Stats": "기기 상태 표시",
-            "Overlay": "오버레이",
-            "Appearance": "모양",
-            "Button Skin": "버튼 스킨",
-            "Custom Skin": "사용자 스킨",
-            "Import Button Images": "버튼 이미지 가져오기",
-            "Haptic Feedback": "햅틱 피드백",
-            "Edit Layout": "레이아웃 편집"
-        ]
-    ]
-
-    private static let commonTranslations: [AppLanguage: [String: String]] = [
-        .simplifiedChinese: [
-            "OSD": "叠加层",
-            "Full Screen": "全屏",
-            "Auto Full Screen": "自动全屏",
-            "Hide Menu Button": "隐藏菜单按钮",
-            "Analog Stick Size": "摇杆大小",
-            "Double-tap empty gameplay space to show the menu button again.": "双击游戏画面空白处可重新显示菜单按钮。",
-            "Menu button shown": "菜单按钮已显示",
-            "Gameplay": "游戏内",
-            "VM is currently running.": "虚拟机正在运行。",
-            "Shut down and start": "关闭并启动",
-            "Compatibility Lab": "兼容性实验室",
-            "Reset ROM": "重置游戏",
-            "Reset ROM?": "重置游戏？",
-            "Restart the current game? Unsaved progress will be lost.": "要重启当前游戏吗？未保存的进度将会丢失。",
-            "Edit Virtual Pad Layout": "编辑虚拟按键布局",
-            "Clear Current Game Cache": "清除此游戏缓存",
-            "Clear Game Cache": "清除游戏缓存",
-            "Cache clear needs a running game.": "需要正在运行的游戏才能清除缓存。",
-            "Game Action": "游戏操作",
-            "Delete Game": "删除游戏",
-            "Delete Game?": "删除游戏？",
-            "Delete ROM": "删除游戏文件",
-            "Delete ROM + Game Data": "删除游戏文件和数据",
-            "Delete Game Data": "删除游戏数据",
-            "Delete Game Data?": "删除游戏数据？",
-            "Delete the selected game file? You can also remove its generated game data at the same time.": "删除所选游戏文件？也可以同时删除生成的游戏数据。",
-            "This clears save states, PNACH files, per-game settings, compatibility overrides, and generated cache for this game. Memory card contents are not deleted.": "这会清除此游戏的即时存档、PNACH 文件、单独设置、兼容性覆盖和缓存。记忆卡内容不会删除。",
-            "Stop this game before deleting it.": "请先停止此游戏再删除。",
-            "Game deleted.": "游戏已删除。",
-            "Could not delete this game file.": "无法删除此游戏文件。",
-            "Game Info": "游戏信息",
-            "Disc": "光盘",
-            "Region": "区域",
-            "Serial": "序列号",
-            "CRC": "CRC",
-            "Format": "格式",
-            "Size": "大小",
-            "File": "文件",
-            "File path unavailable": "文件路径不可用",
-            "Unknown": "未知",
-            "Current Game": "当前游戏",
-            "Current Mode": "当前模式",
-            "Status": "状态",
-            "No Game Active": "没有正在运行的游戏",
-            "Start a game before changing per-game settings.": "请先启动游戏再修改单独游戏设置。",
-            "Per-game settings need a running game.": "需要正在运行的游戏才能修改单独游戏设置。",
-            "Auto Game Presets": "自动游戏预设",
-            "Auto Game Presets applies known safe defaults. Manual flags below are remembered for the current game when a game is running.": "自动游戏预设会应用已知安全默认值。游戏运行时，下方手动标志会保存到当前游戏。",
-            "Start a game to remember presets per title.": "启动游戏后可按标题保存预设。",
-            "Presets": "预设",
-            "Off / Default": "关闭 / 默认",
-            "Use Default / Clear Flags": "使用默认 / 清除标志",
-            "Reset": "重置",
-            "Use this when testing is done or a game behaves worse with compatibility flags enabled.": "测试完成或启用兼容性标志后游戏表现更差时使用此项。",
-            "Custom Advanced Flags": "自定义高级标志",
-            "Manual Compatibility Flags": "手动兼容性标志",
-            "Toggle one or more flags when a game needs compatibility help. Changing any flag switches this game to Custom Advanced Flags.": "当游戏需要兼容性帮助时启用一个或多个标志。更改任意标志会把此游戏切换为自定义高级标志。",
-            "Forget This Game's Override": "忘记此游戏的覆盖设置",
-            "Compatibility preset reset for": "已重置兼容性预设：",
-            "Compatibility preset reset for this game.": "已重置此游戏的兼容性预设。",
-            "Presets are saved for this game and apply on the next boot/reset. Use Off / Default when a preset makes rendering or stability worse.": "预设会保存到此游戏，并在下次启动/重置时生效。如果预设让画面或稳定性变差，请使用关闭/默认。",
-            "Custom compatibility flags saved for": "已保存自定义兼容性标志：",
-            "Compatibility preset set to": "兼容性预设已设为",
-            "saved for this game. Reset or relaunch to apply.": "已保存到此游戏。重置或重新启动后生效。",
-            "Use Per-Game Overrides": "使用单独游戏设置",
-            "Overrides are saved for this game only and apply on the next boot/reset of this title.": "覆盖设置只保存到此游戏，并在下次启动/重置时生效。",
-            "CPU": "CPU",
-            "EE Core": "EE 核心",
-            "ARM64 JIT": "ARM64 JIT",
-            "Interpreter": "解释器",
-            "Interpreter is slower, but can help isolate EE JIT crashes for specific games. Reset/relaunch after changing it.": "解释器较慢，但可用于排查某些游戏的 EE JIT 崩溃。更改后请重置/重启游戏。",
-            "Internal Resolution": "内部渲染分辨率",
-            "Aspect Ratio": "宽高比",
-            "Texture Filtering": "纹理过滤",
-            "Hardware Mipmapping": "硬件 Mipmap",
-            "Blending Accuracy": "混合精度",
-            "Patches & Cheats": "补丁与作弊",
-            "Enable PNACH Cheats": "启用 PNACH 作弊",
-            "GameDB PNACH Patches": "GameDB PNACH 补丁",
-            "GameDB Core Fixes": "GameDB 核心修复",
-            "GameDB Graphics Fixes": "GameDB 图形修复",
-            "Saved for": "已保存到",
-            "Reset or relaunch the game to apply.": "重置或重新启动游戏后生效。",
-            "Per-game overrides cleared.": "已清除单独游戏设置。",
-            "Speed / FPS Target": "速度 / FPS 目标",
-            "Test Controller Rumble": "测试手柄震动",
-            "Change Disc": "换盘",
-            "Eject Disc": "弹出光盘",
-            "No disc images found": "未找到光盘镜像",
-            "Insert Disc (No Reboot)": "插入光盘（不重启）",
-            "Restart With Disc": "用此光盘重启",
-            "Import PNACH / 60 FPS Patch": "导入 PNACH / 60 FPS 补丁",
-            "Save / Load States": "保存 / 读取状态",
-            "Back to Menu": "返回菜单",
-            "Restarting ROM...": "正在重置游戏...",
-            "Now Running": "正在运行",
-            "Stop Emulation": "停止模拟",
-            "Stop Emulation?": "停止模拟？",
-            "This will shut down the running game. All unsaved progress will be lost.": "这会关闭正在运行的游戏，未保存的进度将丢失。",
-            "Resume": "继续",
-            "Download Cover": "下载封面",
-            "Choose Cover": "选择封面",
-            "Remove Cover": "移除封面",
-            "Import Local Covers": "导入本地封面",
-            "Download Missing Covers": "下载缺失封面",
-            "Covers": "封面",
-            "Game Data": "游戏数据",
-            "Library Layout": "游戏列表布局",
-            "Landscape Cover Flow": "横屏封面流",
-            "Show List": "显示列表",
-            "Show Grid": "显示网格",
-            "JIT Status": "JIT 状态",
-            "JIT Enabled": "JIT 已启用",
-            "JIT Access Detected": "已检测到 JIT 访问权限",
-            "JIT Off": "JIT 已关闭",
-            "JIT Not Detected": "未检测到 JIT",
-            "Interpreter Fallback": "解释器回退",
-            "StikDebug Did Not Open": "StikDebug 未打开",
-            "Ready": "就绪",
-            "Check OSD": "检查 OSD",
-            "Manual Open": "手动打开",
-            "Needs StikDebug": "需要 StikDebug",
-            "Open StikDebug": "打开 StikDebug",
-            "StikDebug": "StikDebug",
-            "Auto-open StikDebug": "自动打开 StikDebug",
-            "Access is available. Confirm EE:JIT / IOP:JIT / VU:JIT in the in-game OSD.": "JIT 访问权限可用。请在游戏内 OSD 中确认 EE:JIT / IOP:JIT / VU:JIT。",
-            "Open StikDebug manually, then run the UTM-Dolphin script and relaunch ARMSX2.": "请手动打开 StikDebug，然后运行 UTM-Dolphin 脚本并重新启动 ARMSX2。",
-            "No-JIT fallback is active. Use StikDebug/UTM-Dolphin for dynarec.": "无 JIT 回退已启用。请使用 StikDebug/UTM-Dolphin 启用 dynarec。",
-            "Launch with StikDebug/UTM-Dolphin to enable JIT.": "通过 StikDebug/UTM-Dolphin 启动以启用 JIT。",
-            "ARMSX2 can open StikDebug, but iOS still requires the correct StikDebug/UTM-Dolphin script to attach or relaunch with JIT.": "ARMSX2 可以打开 StikDebug，但 iOS 仍需要正确的 StikDebug/UTM-Dolphin 脚本来附加或用 JIT 重新启动。",
-            "JIT Access means iOS currently allows executable memory. Confirm the real runtime state in-game: the OSD should show EE:JIT, IOP:JIT, and VU:JIT. If StikDebug reports a socket failure, reopen StikDebug and relaunch with the UTM-Dolphin script.": "JIT 访问权限表示 iOS 当前允许可执行内存。请在游戏内确认真实运行状态：OSD 应显示 EE:JIT、IOP:JIT 和 VU:JIT。如果 StikDebug 报告 socket 失败，请重新打开 StikDebug 并用 UTM-Dolphin 脚本重新启动。",
-            "Opens StikDebug automatically when JIT is missing. If the status stays red, launch ARMSX2 from the StikDebug/UTM-Dolphin script.": "JIT 缺失时自动打开 StikDebug。如果状态仍为红色，请从 StikDebug/UTM-Dolphin 脚本启动 ARMSX2。"
-        ],
-        .spanish: [
-            "OSD": "OSD",
-            "Full Screen": "Pantalla completa",
-            "Auto Full Screen": "Pantalla completa automática",
-            "Hide Menu Button": "Ocultar botón de menú",
-            "Analog Stick Size": "Tamaño del stick analógico",
-            "Double-tap empty gameplay space to show the menu button again.": "Toca dos veces un espacio vacío del juego para volver a mostrar el botón de menú.",
-            "Menu button shown": "Botón de menú mostrado",
-            "Gameplay": "Juego",
-            "VM is currently running.": "La VM está en ejecución.",
-            "Shut down and start": "Apagar e iniciar",
-            "Compatibility Lab": "Laboratorio de compatibilidad",
-            "Reset ROM": "Reiniciar juego",
-            "Reset ROM?": "¿Reiniciar juego?",
-            "Restart the current game? Unsaved progress will be lost.": "¿Reiniciar el juego actual? Se perderá el progreso no guardado.",
-            "Edit Virtual Pad Layout": "Editar controles táctiles",
-            "Clear Current Game Cache": "Borrar caché del juego actual",
-            "Clear Game Cache": "Borrar caché del juego",
-            "Cache clear needs a running game.": "Necesitas un juego en ejecución para borrar su caché.",
-            "Game Action": "Acción del juego",
-            "Delete Game": "Eliminar juego",
-            "Delete Game?": "¿Eliminar juego?",
-            "Delete ROM": "Eliminar ROM",
-            "Delete ROM + Game Data": "Eliminar ROM + datos",
-            "Delete Game Data": "Eliminar datos del juego",
-            "Delete Game Data?": "¿Eliminar datos del juego?",
-            "Delete the selected game file? You can also remove its generated game data at the same time.": "¿Eliminar el archivo del juego seleccionado? También puedes borrar sus datos generados.",
-            "This clears save states, PNACH files, per-game settings, compatibility overrides, and generated cache for this game. Memory card contents are not deleted.": "Esto borra estados, PNACH, ajustes por juego, compatibilidad y caché generada. No borra las tarjetas de memoria.",
-            "Stop this game before deleting it.": "Detén este juego antes de eliminarlo.",
-            "Game deleted.": "Juego eliminado.",
-            "Could not delete this game file.": "No se pudo eliminar este juego.",
-            "Game Info": "Información del juego",
-            "Disc": "Disco",
-            "Region": "Región",
-            "Serial": "Serial",
-            "CRC": "CRC",
-            "Format": "Formato",
-            "Size": "Tamaño",
-            "File": "Archivo",
-            "File path unavailable": "Ruta no disponible",
-            "Unknown": "Desconocido",
-            "Current Game": "Juego actual",
-            "Current Mode": "Modo actual",
-            "Status": "Estado",
-            "No Game Active": "No hay juego activo",
-            "Start a game before changing per-game settings.": "Inicia un juego antes de cambiar ajustes por juego.",
-            "Per-game settings need a running game.": "Los ajustes por juego necesitan un juego en ejecución.",
-            "Auto Game Presets": "Presets automáticos",
-            "Auto Game Presets applies known safe defaults. Manual flags below are remembered for the current game when a game is running.": "Los presets automáticos aplican valores seguros conocidos. Las opciones manuales se guardan para el juego actual mientras está en ejecución.",
-            "Start a game to remember presets per title.": "Inicia un juego para recordar presets por título.",
-            "Presets": "Presets",
-            "Off / Default": "Apagado / predeterminado",
-            "Use Default / Clear Flags": "Usar predeterminado / borrar opciones",
-            "Reset": "Reiniciar",
-            "Use this when testing is done or a game behaves worse with compatibility flags enabled.": "Úsalo cuando termines de probar o si un juego empeora con opciones de compatibilidad.",
-            "Custom Advanced Flags": "Opciones avanzadas personalizadas",
-            "Manual Compatibility Flags": "Opciones de compatibilidad manuales",
-            "Toggle one or more flags when a game needs compatibility help. Changing any flag switches this game to Custom Advanced Flags.": "Activa una o más opciones cuando un juego necesite ayuda de compatibilidad. Cambiar cualquiera usa opciones avanzadas personalizadas.",
-            "Forget This Game's Override": "Olvidar ajustes de este juego",
-            "Compatibility preset reset for": "Preset de compatibilidad reiniciado para",
-            "Compatibility preset reset for this game.": "Preset de compatibilidad reiniciado para este juego.",
-            "Presets are saved for this game and apply on the next boot/reset. Use Off / Default when a preset makes rendering or stability worse.": "Los presets se guardan para este juego y se aplican al reiniciar. Usa Apagado/Predeterminado si empeora gráficos o estabilidad.",
-            "Custom compatibility flags saved for": "Opciones de compatibilidad guardadas para",
-            "Compatibility preset set to": "Preset de compatibilidad cambiado a",
-            "saved for this game. Reset or relaunch to apply.": "guardado para este juego. Reinicia para aplicar.",
-            "Use Per-Game Overrides": "Usar ajustes por juego",
-            "Overrides are saved for this game only and apply on the next boot/reset of this title.": "Los ajustes solo se guardan para este juego y se aplican al reiniciar.",
-            "CPU": "CPU",
-            "EE Core": "Núcleo EE",
-            "ARM64 JIT": "ARM64 JIT",
-            "Interpreter": "Intérprete",
-            "Interpreter is slower, but can help isolate EE JIT crashes for specific games. Reset/relaunch after changing it.": "El intérprete es más lento, pero ayuda a aislar fallos del EE JIT. Reinicia tras cambiarlo.",
-            "Internal Resolution": "Resolución interna",
-            "Aspect Ratio": "Relación de aspecto",
-            "Texture Filtering": "Filtrado de texturas",
-            "Hardware Mipmapping": "Mipmapping por hardware",
-            "Blending Accuracy": "Precisión de mezcla",
-            "Patches & Cheats": "Parches y trucos",
-            "Enable PNACH Cheats": "Activar trucos PNACH",
-            "GameDB PNACH Patches": "Parches PNACH de GameDB",
-            "GameDB Core Fixes": "Arreglos del núcleo GameDB",
-            "GameDB Graphics Fixes": "Arreglos gráficos GameDB",
-            "Saved for": "Guardado para",
-            "Reset or relaunch the game to apply.": "Reinicia el juego para aplicar.",
-            "Per-game overrides cleared.": "Ajustes por juego borrados.",
-            "Speed / FPS Target": "Velocidad / FPS objetivo",
-            "Test Controller Rumble": "Probar vibración",
-            "Change Disc": "Cambiar disco",
-            "Eject Disc": "Expulsar disco",
-            "No disc images found": "No se encontraron discos",
-            "Insert Disc (No Reboot)": "Insertar disco sin reiniciar",
-            "Restart With Disc": "Reiniciar con disco",
-            "Import PNACH / 60 FPS Patch": "Importar PNACH / parche 60 FPS",
-            "Save / Load States": "Guardar / cargar estados",
-            "Back to Menu": "Volver al menú",
-            "Restarting ROM...": "Reiniciando juego...",
-            "Now Running": "En ejecución",
-            "Stop Emulation": "Detener emulación",
-            "Stop Emulation?": "¿Detener emulación?",
-            "This will shut down the running game. All unsaved progress will be lost.": "Esto cerrará el juego en ejecución. Se perderá el progreso no guardado.",
-            "Download Cover": "Descargar carátula",
-            "Choose Cover": "Elegir carátula",
-            "Remove Cover": "Quitar carátula",
-            "Import Local Covers": "Importar carátulas locales",
-            "Download Missing Covers": "Descargar carátulas faltantes",
-            "Covers": "Carátulas",
-            "Game Data": "Datos del juego",
-            "Library Layout": "Diseño de biblioteca",
-            "Landscape Cover Flow": "Cover Flow horizontal",
-            "Show List": "Mostrar lista",
-            "Show Grid": "Mostrar cuadrícula",
-            "JIT Status": "Estado JIT",
-            "JIT Enabled": "JIT activado",
-            "JIT Access Detected": "Acceso JIT detectado",
-            "JIT Off": "JIT desactivado",
-            "JIT Not Detected": "JIT no detectado",
-            "Interpreter Fallback": "Intérprete de respaldo",
-            "StikDebug Did Not Open": "StikDebug no se abrió",
-            "Ready": "Listo",
-            "Check OSD": "Revisar OSD",
-            "Manual Open": "Abrir manualmente",
-            "Needs StikDebug": "Requiere StikDebug",
-            "Open StikDebug": "Abrir StikDebug",
-            "StikDebug": "StikDebug",
-            "Auto-open StikDebug": "Abrir StikDebug automáticamente",
-            "Access is available. Confirm EE:JIT / IOP:JIT / VU:JIT in the in-game OSD.": "El acceso está disponible. Confirma EE:JIT / IOP:JIT / VU:JIT en el OSD del juego.",
-            "Open StikDebug manually, then run the UTM-Dolphin script and relaunch ARMSX2.": "Abre StikDebug manualmente, ejecuta el script UTM-Dolphin y vuelve a iniciar ARMSX2.",
-            "No-JIT fallback is active. Use StikDebug/UTM-Dolphin for dynarec.": "El modo sin JIT está activo. Usa StikDebug/UTM-Dolphin para dynarec.",
-            "Launch with StikDebug/UTM-Dolphin to enable JIT.": "Inicia con StikDebug/UTM-Dolphin para activar JIT.",
-            "ARMSX2 can open StikDebug, but iOS still requires the correct StikDebug/UTM-Dolphin script to attach or relaunch with JIT.": "ARMSX2 puede abrir StikDebug, pero iOS aún requiere el script correcto de StikDebug/UTM-Dolphin para adjuntar o relanzar con JIT.",
-            "JIT Access means iOS currently allows executable memory. Confirm the real runtime state in-game: the OSD should show EE:JIT, IOP:JIT, and VU:JIT. If StikDebug reports a socket failure, reopen StikDebug and relaunch with the UTM-Dolphin script.": "Acceso JIT significa que iOS permite memoria ejecutable en este momento. Confirma el estado real en el juego: el OSD debe mostrar EE:JIT, IOP:JIT y VU:JIT. Si StikDebug informa un fallo de socket, vuelve a abrir StikDebug y relanza con el script UTM-Dolphin.",
-            "Opens StikDebug automatically when JIT is missing. If the status stays red, launch ARMSX2 from the StikDebug/UTM-Dolphin script.": "Abre StikDebug automáticamente cuando falta JIT. Si el estado sigue rojo, inicia ARMSX2 desde el script StikDebug/UTM-Dolphin."
-        ],
-        .portuguese: [
-            "OSD": "OSD",
-            "Full Screen": "Tela cheia",
-            "Auto Full Screen": "Tela cheia automática",
-            "Hide Menu Button": "Ocultar botão de menu",
-            "Analog Stick Size": "Tamanho do analógico",
-            "Double-tap empty gameplay space to show the menu button again.": "Toque duas vezes em uma área vazia do jogo para mostrar o botão de menu novamente.",
-            "Menu button shown": "Botão de menu exibido",
-            "Gameplay": "Jogo",
-            "VM is currently running.": "A VM está em execução.",
-            "Shut down and start": "Desligar e iniciar",
-            "Compatibility Lab": "Laboratório de compatibilidade",
-            "Reset ROM": "Reiniciar jogo",
-            "Reset ROM?": "Reiniciar jogo?",
-            "Restart the current game? Unsaved progress will be lost.": "Reiniciar o jogo atual? O progresso não salvo será perdido.",
-            "Edit Virtual Pad Layout": "Editar controles virtuais",
-            "Clear Current Game Cache": "Limpar cache do jogo atual",
-            "Clear Game Cache": "Limpar cache do jogo",
-            "Cache clear needs a running game.": "É preciso ter um jogo em execução para limpar o cache.",
-            "Game Action": "Ação do jogo",
-            "Delete Game": "Excluir jogo",
-            "Delete Game?": "Excluir jogo?",
-            "Delete ROM": "Excluir ROM",
-            "Delete ROM + Game Data": "Excluir ROM + dados",
-            "Delete Game Data": "Excluir dados do jogo",
-            "Delete Game Data?": "Excluir dados do jogo?",
-            "Delete the selected game file? You can also remove its generated game data at the same time.": "Excluir o arquivo do jogo selecionado? Você também pode remover os dados gerados.",
-            "This clears save states, PNACH files, per-game settings, compatibility overrides, and generated cache for this game. Memory card contents are not deleted.": "Isso limpa save states, PNACH, ajustes por jogo, compatibilidade e cache gerado. Cartões de memória não são apagados.",
-            "Stop this game before deleting it.": "Pare este jogo antes de excluí-lo.",
-            "Game deleted.": "Jogo excluído.",
-            "Could not delete this game file.": "Não foi possível excluir este jogo.",
-            "Game Info": "Informações do jogo",
-            "Disc": "Disco",
-            "Region": "Região",
-            "Serial": "Serial",
-            "CRC": "CRC",
-            "Format": "Formato",
-            "Size": "Tamanho",
-            "File": "Arquivo",
-            "File path unavailable": "Caminho indisponível",
-            "Unknown": "Desconhecido",
-            "Current Game": "Jogo atual",
-            "Current Mode": "Modo atual",
-            "Status": "Status",
-            "No Game Active": "Nenhum jogo ativo",
-            "Start a game before changing per-game settings.": "Inicie um jogo antes de alterar ajustes por jogo.",
-            "Per-game settings need a running game.": "Ajustes por jogo precisam de um jogo em execução.",
-            "Auto Game Presets": "Predefinições automáticas",
-            "Auto Game Presets applies known safe defaults. Manual flags below are remembered for the current game when a game is running.": "As predefinições automáticas aplicam padrões seguros conhecidos. As opções manuais abaixo são lembradas para o jogo atual enquanto ele roda.",
-            "Start a game to remember presets per title.": "Inicie um jogo para lembrar predefinições por título.",
-            "Presets": "Predefinições",
-            "Off / Default": "Desligado / padrão",
-            "Use Default / Clear Flags": "Usar padrão / limpar opções",
-            "Reset": "Redefinir",
-            "Use this when testing is done or a game behaves worse with compatibility flags enabled.": "Use quando os testes terminarem ou quando um jogo piorar com opções de compatibilidade.",
-            "Custom Advanced Flags": "Opções avançadas personalizadas",
-            "Manual Compatibility Flags": "Opções manuais de compatibilidade",
-            "Toggle one or more flags when a game needs compatibility help. Changing any flag switches this game to Custom Advanced Flags.": "Ative uma ou mais opções quando um jogo precisar de ajuda de compatibilidade. Alterar qualquer opção muda para opções avançadas personalizadas.",
-            "Forget This Game's Override": "Esquecer ajustes deste jogo",
-            "Compatibility preset reset for": "Predefinição de compatibilidade redefinida para",
-            "Compatibility preset reset for this game.": "Predefinição de compatibilidade redefinida para este jogo.",
-            "Presets are saved for this game and apply on the next boot/reset. Use Off / Default when a preset makes rendering or stability worse.": "As predefinições são salvas por jogo e aplicadas ao reiniciar. Use Desligado/Padrão se piorar gráficos ou estabilidade.",
-            "Custom compatibility flags saved for": "Opções de compatibilidade salvas para",
-            "Compatibility preset set to": "Predefinição de compatibilidade definida como",
-            "saved for this game. Reset or relaunch to apply.": "salvo para este jogo. Reinicie para aplicar.",
-            "Use Per-Game Overrides": "Usar ajustes por jogo",
-            "Overrides are saved for this game only and apply on the next boot/reset of this title.": "Os ajustes são salvos apenas para este jogo e aplicam ao reiniciar.",
-            "CPU": "CPU",
-            "EE Core": "Núcleo EE",
-            "ARM64 JIT": "ARM64 JIT",
-            "Interpreter": "Interpretador",
-            "Interpreter is slower, but can help isolate EE JIT crashes for specific games. Reset/relaunch after changing it.": "O interpretador é mais lento, mas ajuda a isolar falhas do EE JIT. Reinicie após alterar.",
-            "Internal Resolution": "Resolução interna",
-            "Aspect Ratio": "Proporção de tela",
-            "Texture Filtering": "Filtragem de textura",
-            "Hardware Mipmapping": "Mipmapping por hardware",
-            "Blending Accuracy": "Precisão de mistura",
-            "Patches & Cheats": "Patches e cheats",
-            "Enable PNACH Cheats": "Ativar cheats PNACH",
-            "GameDB PNACH Patches": "Patches PNACH do GameDB",
-            "GameDB Core Fixes": "Correções de núcleo GameDB",
-            "GameDB Graphics Fixes": "Correções gráficas GameDB",
-            "Saved for": "Salvo para",
-            "Reset or relaunch the game to apply.": "Reinicie o jogo para aplicar.",
-            "Per-game overrides cleared.": "Ajustes por jogo limpos.",
-            "Speed / FPS Target": "Velocidade / FPS alvo",
-            "Test Controller Rumble": "Testar vibração",
-            "Change Disc": "Trocar disco",
-            "Eject Disc": "Ejetar disco",
-            "No disc images found": "Nenhuma imagem de disco encontrada",
-            "Insert Disc (No Reboot)": "Inserir disco sem reiniciar",
-            "Restart With Disc": "Reiniciar com disco",
-            "Import PNACH / 60 FPS Patch": "Importar PNACH / patch 60 FPS",
-            "Save / Load States": "Salvar / carregar estados",
-            "Back to Menu": "Voltar ao menu",
-            "Restarting ROM...": "Reiniciando jogo...",
-            "Now Running": "Em execução",
-            "Stop Emulation": "Parar emulação",
-            "Stop Emulation?": "Parar emulação?",
-            "This will shut down the running game. All unsaved progress will be lost.": "Isso encerrará o jogo em execução. O progresso não salvo será perdido.",
-            "Download Cover": "Baixar capa",
-            "Choose Cover": "Escolher capa",
-            "Remove Cover": "Remover capa",
-            "Import Local Covers": "Importar capas locais",
-            "Download Missing Covers": "Baixar capas ausentes",
-            "Covers": "Capas",
-            "Game Data": "Dados do jogo",
-            "Library Layout": "Layout da biblioteca",
-            "Landscape Cover Flow": "Cover Flow em paisagem",
-            "Show List": "Mostrar lista",
-            "Show Grid": "Mostrar grade",
-            "JIT Status": "Estado do JIT",
-            "JIT Enabled": "JIT ativado",
-            "JIT Access Detected": "Acesso JIT detectado",
-            "JIT Off": "JIT desligado",
-            "JIT Not Detected": "JIT não detectado",
-            "Interpreter Fallback": "Interpretador de fallback",
-            "StikDebug Did Not Open": "StikDebug não abriu",
-            "Ready": "Pronto",
-            "Check OSD": "Verificar OSD",
-            "Manual Open": "Abrir manualmente",
-            "Needs StikDebug": "Precisa do StikDebug",
-            "Open StikDebug": "Abrir StikDebug",
-            "StikDebug": "StikDebug",
-            "Auto-open StikDebug": "Abrir StikDebug automaticamente",
-            "Access is available. Confirm EE:JIT / IOP:JIT / VU:JIT in the in-game OSD.": "O acesso está disponível. Confirme EE:JIT / IOP:JIT / VU:JIT no OSD dentro do jogo.",
-            "Open StikDebug manually, then run the UTM-Dolphin script and relaunch ARMSX2.": "Abra o StikDebug manualmente, execute o script UTM-Dolphin e reinicie o ARMSX2.",
-            "No-JIT fallback is active. Use StikDebug/UTM-Dolphin for dynarec.": "O modo sem JIT está ativo. Use StikDebug/UTM-Dolphin para dynarec.",
-            "Launch with StikDebug/UTM-Dolphin to enable JIT.": "Inicie com StikDebug/UTM-Dolphin para ativar o JIT.",
-            "ARMSX2 can open StikDebug, but iOS still requires the correct StikDebug/UTM-Dolphin script to attach or relaunch with JIT.": "O ARMSX2 pode abrir o StikDebug, mas o iOS ainda exige o script correto do StikDebug/UTM-Dolphin para anexar ou reiniciar com JIT.",
-            "JIT Access means iOS currently allows executable memory. Confirm the real runtime state in-game: the OSD should show EE:JIT, IOP:JIT, and VU:JIT. If StikDebug reports a socket failure, reopen StikDebug and relaunch with the UTM-Dolphin script.": "Acesso JIT significa que o iOS permite memória executável no momento. Confirme o estado real no jogo: o OSD deve mostrar EE:JIT, IOP:JIT e VU:JIT. Se o StikDebug informar falha de socket, reabra o StikDebug e reinicie com o script UTM-Dolphin.",
-            "Opens StikDebug automatically when JIT is missing. If the status stays red, launch ARMSX2 from the StikDebug/UTM-Dolphin script.": "Abre o StikDebug automaticamente quando o JIT está ausente. Se o estado continuar vermelho, inicie o ARMSX2 pelo script StikDebug/UTM-Dolphin."
-        ]
-    ]
-
-    private static let uiSupplementTranslations: [AppLanguage: [String: String]] = [
-        .simplifiedChinese: [
-            "Configure multitap for 3-4 local controllers.": "配置 3-4 个本地手柄的 Multitap。",
-            "Button Mapping": "按键映射",
-            "Tap a row, then press a button on your controller to assign it. L2/R2 are analog triggers (not remappable).": "点击一行，然后按下手柄按键进行绑定。L2/R2 是模拟扳机，不能重映射。",
-            "Controller rumble test sent.": "已发送手柄震动测试。",
-            "Reset to Default": "恢复默认",
-            "Tools": "工具",
-            "Press a button...": "请按一个按键...",
-            "D-Pad Up": "十字键 上",
-            "D-Pad Down": "十字键 下",
-            "D-Pad Left": "十字键 左",
-            "D-Pad Right": "十字键 右",
-            "Cross": "叉",
-            "Circle": "圆",
-            "Square": "方块",
-            "Triangle": "三角",
-            "Start": "开始",
-            "Select": "选择",
-            "A / Cross": "A / 叉",
-            "B / Circle": "B / 圆",
-            "X / Square": "X / 方块",
-            "Y / Triangle": "Y / 三角",
-            "Share / Back": "分享 / 返回",
-            "Guide / PS": "主页 / PS",
-            "Options / Start": "选项 / 开始",
-            "L-Stick Press": "左摇杆按下",
-            "R-Stick Press": "右摇杆按下",
-            "L-Shoulder": "左肩键",
-            "R-Shoulder": "右肩键",
-            "Misc / Share": "其他 / 分享",
-            "Touchpad": "触控板",
-            "Multitap Mode": "Multitap 模式",
-            "Auto (3+ Controllers)": "自动（3 个以上手柄）",
-            "Auto": "自动",
-            "Disabled": "关闭",
-            "Port 1 Multitap": "端口 1 Multitap",
-            "Port 2 Multitap": "端口 2 Multitap",
-            "Port 1 + Port 2 Multitap": "端口 1 + 端口 2 Multitap",
-            "Multitap": "Multitap",
-            "Auto enables Port 1 multitap when 3 or more controllers are detected before boot. Manual modes take effect on the next boot/reset.": "启动前检测到 3 个或更多手柄时，自动启用端口 1 Multitap。手动模式在下次启动/重置时生效。",
-            "Controller Mapping": "手柄映射",
-            "Disabled maps controllers 1-2 to normal PS2 ports. Port 1 Multitap maps controllers 1-4 to 1A/1B/1C/1D. Port 2 Multitap keeps controller 1 on Port 1 and maps controllers 2-4 to Port 2 multitap slots.": "关闭时手柄 1-2 映射到普通 PS2 端口。端口 1 Multitap 将手柄 1-4 映射到 1A/1B/1C/1D。端口 2 Multitap 保留手柄 1 在端口 1，并将手柄 2-4 映射到端口 2 的 Multitap 插槽。",
-            "Enable RetroAchievements": "启用 RetroAchievements",
-            "Client": "客户端",
-            "Active": "已启用",
-            "Inactive": "未启用",
-            "Uses the same achievements core as ARMSX2 Android. Login tokens are stored in ARMSX2's local config; passwords are not stored by this screen.": "使用与 ARMSX2 Android 相同的成就核心。登录令牌保存在 ARMSX2 本地配置中；此界面不会保存密码。",
-            "User": "用户",
-            "Points": "点数",
-            "Messages": "消息",
-            "unread": "未读",
-            "Log Out": "退出登录",
-            "Logging In...": "正在登录...",
-            "Log In": "登录",
-            "Modes": "模式",
-            "Hardcore Mode": "硬核模式",
-            "Achievement Notifications": "成就通知",
-            "Leaderboard Notifications": "排行榜通知",
-            "In-Game Overlays": "游戏内叠加层",
-            "Hardcore is enforced by the core and can restrict cheats, save states, and other non-hardcore features while active.": "硬核模式由核心强制执行，启用时可能限制作弊、即时存档和其他非硬核功能。",
-            "Title": "标题",
-            "Game ID": "游戏 ID",
-            "Unknown Game": "未知游戏",
-            "Achievements": "成就",
-            "Loaded": "已加载",
-            "None found": "未找到",
-            "Leaderboards": "排行榜",
-            "Available": "可用",
-            "Rich Presence": "状态显示",
-            "Boot a game while RetroAchievements is enabled to see game progress here.": "启用 RetroAchievements 后启动游戏，可在此查看进度。",
-            "RetroAchievements Login": "RetroAchievements 登录",
-            "Username": "用户名",
-            "Password": "密码",
-            "Use your RetroAchievements account credentials.": "使用你的 RetroAchievements 账户凭据。",
-            "Missing Login": "缺少登录信息",
-            "Enter both your RetroAchievements username and password.": "请输入 RetroAchievements 用户名和密码。",
-            "Logged In": "已登录",
-            "Login Failed": "登录失败",
-            "Not logged in": "未登录",
-            "Directory": "目录",
-            "Assigned Cards": "已分配记忆卡",
-            "Slot 1": "插槽 1",
-            "Slot 2": "插槽 2",
-            "Unplugged": "未插入",
-            "Slot changes take effect on the next VM boot.": "插槽更改会在下次虚拟机启动时生效。",
-            "Card name": "记忆卡名称",
-            "Create Folder Card": "创建文件夹记忆卡",
-            "Create Card": "创建记忆卡",
-            "File cards support 8 MB, 16 MB, 32 MB, and 64 MB. Folder cards match the ARMSX2/PCSX2 folder-card behavior and are useful for game-specific saves.": "文件记忆卡支持 8 MB、16 MB、32 MB 和 64 MB。文件夹记忆卡符合 ARMSX2/PCSX2 行为，适合按游戏保存。",
-            "No cards found.": "未找到记忆卡。",
-            "Memory card created.": "记忆卡已创建。",
-            "Could not create memory card. Check the name, size, or whether it already exists.": "无法创建记忆卡。请检查名称、大小或是否已存在。",
-            "PS2 HDD": "PS2 硬盘",
-            "Enable DEV9 Virtual HDD": "启用 DEV9 虚拟硬盘",
-            "Image": "镜像",
-            "Use Default HDD Image": "使用默认硬盘镜像",
-            "Matches ARMSX2 Android's DEV9 HDD option. Requires a VM restart.": "匹配 ARMSX2 Android 的 DEV9 HDD 选项。需要重启虚拟机。",
-            "Online / Ethernet": "在线 / 以太网",
-            "Enable DEV9 Ethernet": "启用 DEV9 以太网",
-            "Mode": "模式",
-            "Sockets": "Sockets",
-            "Adapter": "适配器",
-            "Intercept DHCP": "拦截 DHCP",
-            "Log DHCP": "记录 DHCP",
-            "Log DNS": "记录 DNS",
-            "Sockets is the iOS-safe DEV9 Ethernet mode exposed here. PCAP bridged/switched modes are compiled out of this iOS build, so they are intentionally not selectable until a real iOS backend exists.": "Sockets 是这里暴露的 iOS 安全 DEV9 以太网模式。PCAP 桥接/交换模式已从此 iOS 构建中移除，因此在真正的 iOS 后端出现前不可选择。",
-            "Tester Notes": "测试说明",
-            "Games still need their in-game PS2 network setup. After changing DEV9 settings, use Reset ROM or restart the VM before testing.": "游戏仍需要在游戏内完成 PS2 网络设置。更改 DEV9 设置后，请先重置游戏或重启虚拟机再测试。",
-            "CPU Recompiler": "CPU 重编译器",
-            "Changes take effect on next VM boot.": "更改会在下次虚拟机启动时生效。",
-            "Boot": "启动",
-            "Memory": "内存",
-            "Frame Limiter": "帧率限制",
-            "FPS Target": "FPS 目标",
-            "Speed Target": "速度目标",
-            "Unlocked": "不限制",
-            "NTSC Base Rate": "NTSC 基准帧率",
-            "PAL Base Rate": "PAL 基准帧率",
-            "FPS Target maps to PCSX2 Normal Speed: 60 FPS is normal NTSC timing, 30 FPS is about 50% speed, and higher values fast-forward. Turning the limiter OFF unlocks speed and can increase heat and battery drain.": "FPS 目标映射到 PCSX2 正常速度：60 FPS 为正常 NTSC，30 FPS 约为 50% 速度，更高值为快进。关闭限制会解锁速度并增加发热和耗电。",
-            "Use VU1 Interpreter Preset": "使用 VU1 解释器预设",
-            "Use Full Interpreter Preset": "使用完整解释器预设",
-            "GameDB Automatic Fixes": "GameDB 自动修复",
-            "Widescreen Patches": "宽屏补丁",
-            "No-Interlacing Patches": "去隔行补丁",
-            "GameDB Core Fixes covers timing, clamps, and gamefixes. GameDB Graphics Fixes covers renderer-specific hardware fixes; turn it off globally or per-game if a title looks worse on Metal. PNACH cheats and 60 FPS patches can be imported from the in-game quick menu or from a game's long-press menu.": "GameDB 核心修复涵盖时序、钳位和游戏修复。GameDB 图形修复涵盖渲染器相关硬件修复；如果某游戏在 Metal 下更差，可全局或按游戏关闭。PNACH 作弊和 60 FPS 补丁可从游戏内快捷菜单或游戏长按菜单导入。",
-            "EE Cycle Rate": "EE 周期率",
-            "0 = Default. Negative = underclock (stable). Positive = overclock (fast but risky).": "0 = 默认。负数 = 降频（稳定）。正数 = 超频（更快但有风险）。",
-            "Fast CDVD": "快速 CDVD",
-            "Wait Loop Detection": "等待循环检测",
-            "INTC Stat Hack": "INTC 状态 Hack",
-            "VU1 Instant and MTVU are independent now. MTVU can help some games, but keep it off unless a game specifically benefits on iOS.": "VU1 Instant 和 MTVU 现在互相独立。MTVU 可帮助部分游戏，但除非某游戏在 iOS 上明确受益，否则建议关闭。",
-            "Speedhacks": "速度 Hack",
-            "Reset Emulator to Defaults": "重置模拟器设置",
-            "Usage": "用量",
-            "App Cache": "应用缓存",
-            "Texture Dumps": "纹理转储",
-            "Diagnostic Logs": "诊断日志",
-            "State Safety Backups": "即时存档安全备份",
-            "Generated Total": "生成文件总计",
-            "Refresh Storage Usage": "刷新存储用量",
-            "Cleanup": "清理",
-            "Clear App Cache": "清除应用缓存",
-            "Clear Texture Dumps": "清除纹理转储",
-            "Clear Diagnostic Logs": "清除诊断日志",
-            "Clear All Generated Files": "清除所有生成文件",
-            "This never removes imported games, BIOS files, save states, memory cards, covers, settings, PNACH files, or replacement texture packs.": "这不会删除已导入游戏、BIOS、即时存档、记忆卡、封面、设置、PNACH 文件或替换纹理包。",
-            "Renderer": "渲染器",
-            "Upscaling": "升频渲染",
-            "Filtering": "过滤",
-            "Display": "显示",
-            "Quality": "质量",
-            "Texture Replacement": "纹理替换",
-            "Texture Dumping": "纹理转储",
-            "VSync": "垂直同步",
-            "Reset Graphics to Defaults": "重置图形设置",
-            "Load Replacement Textures": "加载替换纹理",
-            "Dump Replaceable Textures": "转储可替换纹理",
-            "Dump Mipmaps": "转储 Mipmap",
-            "Dump During FMV": "FMV 中转储",
-            "Dump Direct Textures": "转储直接纹理",
-            "Dump Palette Textures": "转储调色板纹理",
-            "Queue Size": "队列大小",
-            "Higher values reduce frame drops but increase latency.": "较高值可减少掉帧，但会增加延迟。"
-        ],
-        .spanish: [
-            "Configure multitap for 3-4 local controllers.": "Configura multitap para 3-4 mandos locales.",
-            "Button Mapping": "Asignación de botones",
-            "Tap a row, then press a button on your controller to assign it. L2/R2 are analog triggers (not remappable).": "Toca una fila y pulsa un botón del mando para asignarlo. L2/R2 son gatillos analógicos y no se pueden reasignar.",
-            "Controller rumble test sent.": "Prueba de vibración enviada.",
-            "Reset to Default": "Restablecer valores",
-            "Tools": "Herramientas",
-            "Press a button...": "Pulsa un botón...",
-            "D-Pad Up": "Cruceta arriba",
-            "D-Pad Down": "Cruceta abajo",
-            "D-Pad Left": "Cruceta izquierda",
-            "D-Pad Right": "Cruceta derecha",
-            "Cross": "Cruz",
-            "Circle": "Círculo",
-            "Square": "Cuadrado",
-            "Triangle": "Triángulo",
-            "Start": "Inicio",
-            "Select": "Select",
-            "A / Cross": "A / Cruz",
-            "B / Circle": "B / Círculo",
-            "X / Square": "X / Cuadrado",
-            "Y / Triangle": "Y / Triángulo",
-            "Share / Back": "Compartir / Atrás",
-            "Guide / PS": "Guía / PS",
-            "Options / Start": "Opciones / Inicio",
-            "L-Stick Press": "Pulsar stick izquierdo",
-            "R-Stick Press": "Pulsar stick derecho",
-            "L-Shoulder": "Botón L",
-            "R-Shoulder": "Botón R",
-            "Misc / Share": "Varios / Compartir",
-            "Touchpad": "Panel táctil",
-            "Multitap Mode": "Modo multitap",
-            "Auto (3+ Controllers)": "Auto (3+ mandos)",
-            "Auto": "Auto",
-            "Disabled": "Desactivado",
-            "Port 1 Multitap": "Multitap puerto 1",
-            "Port 2 Multitap": "Multitap puerto 2",
-            "Port 1 + Port 2 Multitap": "Multitap puerto 1 + 2",
-            "Multitap": "Multitap",
-            "Auto enables Port 1 multitap when 3 or more controllers are detected before boot. Manual modes take effect on the next boot/reset.": "Auto activa multitap en el puerto 1 si detecta 3 o más mandos antes de iniciar. Los modos manuales se aplican al siguiente arranque/reinicio.",
-            "Controller Mapping": "Asignación de mandos",
-            "Disabled maps controllers 1-2 to normal PS2 ports. Port 1 Multitap maps controllers 1-4 to 1A/1B/1C/1D. Port 2 Multitap keeps controller 1 on Port 1 and maps controllers 2-4 to Port 2 multitap slots.": "Desactivado asigna los mandos 1-2 a puertos normales de PS2. Multitap puerto 1 asigna mandos 1-4 a 1A/1B/1C/1D. Multitap puerto 2 deja el mando 1 en puerto 1 y asigna mandos 2-4 al multitap del puerto 2.",
-            "Enable RetroAchievements": "Activar RetroAchievements",
-            "Client": "Cliente",
-            "Active": "Activo",
-            "Inactive": "Inactivo",
-            "Uses the same achievements core as ARMSX2 Android. Login tokens are stored in ARMSX2's local config; passwords are not stored by this screen.": "Usa el mismo núcleo de logros que ARMSX2 Android. Los tokens se guardan en la configuración local; esta pantalla no guarda contraseñas.",
-            "User": "Usuario",
-            "Points": "Puntos",
-            "Messages": "Mensajes",
-            "unread": "sin leer",
-            "Log Out": "Cerrar sesión",
-            "Logging In...": "Iniciando sesión...",
-            "Log In": "Iniciar sesión",
-            "Modes": "Modos",
-            "Hardcore Mode": "Modo hardcore",
-            "Achievement Notifications": "Notificaciones de logros",
-            "Leaderboard Notifications": "Notificaciones de marcadores",
-            "In-Game Overlays": "Superposiciones en juego",
-            "Hardcore is enforced by the core and can restrict cheats, save states, and other non-hardcore features while active.": "Hardcore lo aplica el núcleo y puede restringir trucos, estados guardados y otras funciones no hardcore.",
-            "Title": "Título",
-            "Game ID": "ID del juego",
-            "Unknown Game": "Juego desconocido",
-            "Achievements": "Logros",
-            "Loaded": "Cargados",
-            "None found": "No encontrados",
-            "Leaderboards": "Marcadores",
-            "Available": "Disponibles",
-            "Rich Presence": "Presencia",
-            "Boot a game while RetroAchievements is enabled to see game progress here.": "Inicia un juego con RetroAchievements activado para ver el progreso aquí.",
-            "RetroAchievements Login": "Inicio de sesión de RetroAchievements",
-            "Username": "Usuario",
-            "Password": "Contraseña",
-            "Use your RetroAchievements account credentials.": "Usa tus credenciales de RetroAchievements.",
-            "Missing Login": "Faltan datos",
-            "Enter both your RetroAchievements username and password.": "Introduce usuario y contraseña de RetroAchievements.",
-            "Logged In": "Sesión iniciada",
-            "Login Failed": "Error al iniciar sesión",
-            "Not logged in": "Sin sesión",
-            "Directory": "Directorio",
-            "Assigned Cards": "Tarjetas asignadas",
-            "Slot 1": "Ranura 1",
-            "Slot 2": "Ranura 2",
-            "Unplugged": "Desconectada",
-            "Slot changes take effect on the next VM boot.": "Los cambios de ranura se aplican al siguiente arranque de la VM.",
-            "Card name": "Nombre de tarjeta",
-            "Create Folder Card": "Crear tarjeta de carpeta",
-            "Create Card": "Crear tarjeta",
-            "File cards support 8 MB, 16 MB, 32 MB, and 64 MB. Folder cards match the ARMSX2/PCSX2 folder-card behavior and are useful for game-specific saves.": "Las tarjetas de archivo soportan 8, 16, 32 y 64 MB. Las tarjetas de carpeta imitan ARMSX2/PCSX2 y sirven para partidas por juego.",
-            "No cards found.": "No se encontraron tarjetas.",
-            "Memory card created.": "Tarjeta creada.",
-            "Could not create memory card. Check the name, size, or whether it already exists.": "No se pudo crear la tarjeta. Revisa el nombre, tamaño o si ya existe.",
-            "PS2 HDD": "Disco duro PS2",
-            "Enable DEV9 Virtual HDD": "Activar HDD virtual DEV9",
-            "Image": "Imagen",
-            "Use Default HDD Image": "Usar imagen HDD predeterminada",
-            "Matches ARMSX2 Android's DEV9 HDD option. Requires a VM restart.": "Coincide con la opción DEV9 HDD de ARMSX2 Android. Requiere reiniciar la VM.",
-            "Online / Ethernet": "Online / Ethernet",
-            "Enable DEV9 Ethernet": "Activar Ethernet DEV9",
-            "Mode": "Modo",
-            "Sockets": "Sockets",
-            "Adapter": "Adaptador",
-            "Intercept DHCP": "Interceptar DHCP",
-            "Log DHCP": "Registrar DHCP",
-            "Log DNS": "Registrar DNS",
-            "Sockets is the iOS-safe DEV9 Ethernet mode exposed here. PCAP bridged/switched modes are compiled out of this iOS build, so they are intentionally not selectable until a real iOS backend exists.": "Sockets es el modo DEV9 seguro para iOS. Los modos PCAP puente/conmutado están excluidos de esta compilación hasta que exista un backend real de iOS.",
-            "Tester Notes": "Notas para testers",
-            "Games still need their in-game PS2 network setup. After changing DEV9 settings, use Reset ROM or restart the VM before testing.": "Los juegos aún necesitan configurar la red PS2 dentro del juego. Tras cambiar DEV9, usa Reiniciar juego o reinicia la VM antes de probar.",
-            "CPU Recompiler": "Recompilador CPU",
-            "Changes take effect on next VM boot.": "Los cambios se aplican al siguiente arranque de la VM.",
-            "Boot": "Arranque",
-            "Memory": "Memoria",
-            "Frame Limiter": "Limitador de FPS",
-            "FPS Target": "FPS objetivo",
-            "Speed Target": "Velocidad objetivo",
-            "Unlocked": "Sin límite",
-            "NTSC Base Rate": "Frecuencia base NTSC",
-            "PAL Base Rate": "Frecuencia base PAL",
-            "FPS Target maps to PCSX2 Normal Speed: 60 FPS is normal NTSC timing, 30 FPS is about 50% speed, and higher values fast-forward. Turning the limiter OFF unlocks speed and can increase heat and battery drain.": "FPS objetivo se asigna a la velocidad normal de PCSX2: 60 FPS es NTSC normal, 30 FPS es ~50%, y valores mayores aceleran. Apagar el limitador desbloquea la velocidad y puede aumentar calor y batería.",
-            "Use VU1 Interpreter Preset": "Usar preset intérprete VU1",
-            "Use Full Interpreter Preset": "Usar preset intérprete completo",
-            "GameDB Automatic Fixes": "Arreglos automáticos GameDB",
-            "Widescreen Patches": "Parches widescreen",
-            "No-Interlacing Patches": "Parches sin entrelazado",
-            "GameDB Core Fixes covers timing, clamps, and gamefixes. GameDB Graphics Fixes covers renderer-specific hardware fixes; turn it off globally or per-game if a title looks worse on Metal. PNACH cheats and 60 FPS patches can be imported from the in-game quick menu or from a game's long-press menu.": "GameDB Core cubre timing, clamps y arreglos. GameDB Graphics cubre fixes del renderer; apágalo globalmente o por juego si un título se ve peor en Metal. Los PNACH y parches 60 FPS se importan desde el menú rápido o el menú del juego.",
-            "EE Cycle Rate": "Ciclo EE",
-            "0 = Default. Negative = underclock (stable). Positive = overclock (fast but risky).": "0 = predeterminado. Negativo = underclock estable. Positivo = overclock rápido pero riesgoso.",
-            "Fast CDVD": "CDVD rápido",
-            "Wait Loop Detection": "Detección de espera",
-            "INTC Stat Hack": "Hack INTC Stat",
-            "VU1 Instant and MTVU are independent now. MTVU can help some games, but keep it off unless a game specifically benefits on iOS.": "VU1 Instant y MTVU ahora son independientes. MTVU puede ayudar en algunos juegos, pero déjalo apagado salvo que un juego se beneficie en iOS.",
-            "Speedhacks": "Speedhacks",
-            "Reset Emulator to Defaults": "Restablecer emulador",
-            "Usage": "Uso",
-            "App Cache": "Caché de la app",
-            "Texture Dumps": "Volcados de texturas",
-            "Diagnostic Logs": "Registros de diagnóstico",
-            "State Safety Backups": "Copias de seguridad de estados",
-            "Generated Total": "Total generado",
-            "Refresh Storage Usage": "Actualizar uso de almacenamiento",
-            "Cleanup": "Limpieza",
-            "Clear App Cache": "Borrar caché de la app",
-            "Clear Texture Dumps": "Borrar volcados de texturas",
-            "Clear Diagnostic Logs": "Borrar registros",
-            "Clear All Generated Files": "Borrar todo lo generado",
-            "This never removes imported games, BIOS files, save states, memory cards, covers, settings, PNACH files, or replacement texture packs.": "Nunca elimina juegos importados, BIOS, estados, tarjetas, carátulas, ajustes, PNACH ni paquetes de texturas.",
-            "Renderer": "Renderizador",
-            "Upscaling": "Escalado",
-            "Filtering": "Filtrado",
-            "Display": "Pantalla",
-            "Quality": "Calidad",
-            "Texture Replacement": "Reemplazo de texturas",
-            "Texture Dumping": "Volcado de texturas",
-            "VSync": "VSync",
-            "Reset Graphics to Defaults": "Restablecer gráficos",
-            "Load Replacement Textures": "Cargar texturas de reemplazo",
-            "Dump Replaceable Textures": "Volcar texturas reemplazables",
-            "Dump Mipmaps": "Volcar mipmaps",
-            "Dump During FMV": "Volcar durante FMV",
-            "Dump Direct Textures": "Volcar texturas directas",
-            "Dump Palette Textures": "Volcar texturas de paleta",
-            "Queue Size": "Tamaño de cola",
-            "Higher values reduce frame drops but increase latency.": "Valores más altos reducen caídas de frames, pero aumentan la latencia."
-        ],
-        .portuguese: [
-            "Configure multitap for 3-4 local controllers.": "Configure multitap para 3-4 controles locais.",
-            "Button Mapping": "Mapeamento de botões",
-            "Tap a row, then press a button on your controller to assign it. L2/R2 are analog triggers (not remappable).": "Toque em uma linha e pressione um botão no controle para atribuir. L2/R2 são gatilhos analógicos e não podem ser remapeados.",
-            "Controller rumble test sent.": "Teste de vibração enviado.",
-            "Reset to Default": "Restaurar padrão",
-            "Tools": "Ferramentas",
-            "Press a button...": "Pressione um botão...",
-            "D-Pad Up": "Direcional cima",
-            "D-Pad Down": "Direcional baixo",
-            "D-Pad Left": "Direcional esquerda",
-            "D-Pad Right": "Direcional direita",
-            "Cross": "X",
-            "Circle": "Círculo",
-            "Square": "Quadrado",
-            "Triangle": "Triângulo",
-            "Start": "Start",
-            "Select": "Select",
-            "A / Cross": "A / X",
-            "B / Circle": "B / Círculo",
-            "X / Square": "X / Quadrado",
-            "Y / Triangle": "Y / Triângulo",
-            "Share / Back": "Compartilhar / Voltar",
-            "Guide / PS": "Guia / PS",
-            "Options / Start": "Opções / Start",
-            "L-Stick Press": "Pressionar analógico esquerdo",
-            "R-Stick Press": "Pressionar analógico direito",
-            "L-Shoulder": "Botão L",
-            "R-Shoulder": "Botão R",
-            "Misc / Share": "Extra / Compartilhar",
-            "Touchpad": "Touchpad",
-            "Multitap Mode": "Modo multitap",
-            "Auto (3+ Controllers)": "Auto (3+ controles)",
-            "Auto": "Auto",
-            "Disabled": "Desativado",
-            "Port 1 Multitap": "Multitap porta 1",
-            "Port 2 Multitap": "Multitap porta 2",
-            "Port 1 + Port 2 Multitap": "Multitap porta 1 + 2",
-            "Multitap": "Multitap",
-            "Auto enables Port 1 multitap when 3 or more controllers are detected before boot. Manual modes take effect on the next boot/reset.": "Auto ativa multitap na porta 1 quando 3 ou mais controles são detectados antes do boot. Modos manuais valem no próximo boot/reset.",
-            "Controller Mapping": "Mapeamento de controles",
-            "Disabled maps controllers 1-2 to normal PS2 ports. Port 1 Multitap maps controllers 1-4 to 1A/1B/1C/1D. Port 2 Multitap keeps controller 1 on Port 1 and maps controllers 2-4 to Port 2 multitap slots.": "Desativado mapeia controles 1-2 para portas normais do PS2. Multitap porta 1 mapeia controles 1-4 para 1A/1B/1C/1D. Multitap porta 2 mantém o controle 1 na porta 1 e mapeia controles 2-4 para a porta 2.",
-            "Enable RetroAchievements": "Ativar RetroAchievements",
-            "Client": "Cliente",
-            "Active": "Ativo",
-            "Inactive": "Inativo",
-            "Uses the same achievements core as ARMSX2 Android. Login tokens are stored in ARMSX2's local config; passwords are not stored by this screen.": "Usa o mesmo núcleo de conquistas do ARMSX2 Android. Tokens ficam na configuração local; esta tela não salva senhas.",
-            "User": "Usuário",
-            "Points": "Pontos",
-            "Messages": "Mensagens",
-            "unread": "não lidas",
-            "Log Out": "Sair",
-            "Logging In...": "Entrando...",
-            "Log In": "Entrar",
-            "Modes": "Modos",
-            "Hardcore Mode": "Modo hardcore",
-            "Achievement Notifications": "Notificações de conquistas",
-            "Leaderboard Notifications": "Notificações de placares",
-            "In-Game Overlays": "Overlays no jogo",
-            "Hardcore is enforced by the core and can restrict cheats, save states, and other non-hardcore features while active.": "Hardcore é aplicado pelo núcleo e pode restringir cheats, save states e recursos não-hardcore.",
-            "Title": "Título",
-            "Game ID": "ID do jogo",
-            "Unknown Game": "Jogo desconhecido",
-            "Achievements": "Conquistas",
-            "Loaded": "Carregadas",
-            "None found": "Nada encontrado",
-            "Leaderboards": "Placares",
-            "Available": "Disponíveis",
-            "Rich Presence": "Presença",
-            "Boot a game while RetroAchievements is enabled to see game progress here.": "Inicie um jogo com RetroAchievements ativado para ver o progresso aqui.",
-            "RetroAchievements Login": "Login RetroAchievements",
-            "Username": "Usuário",
-            "Password": "Senha",
-            "Use your RetroAchievements account credentials.": "Use sua conta do RetroAchievements.",
-            "Missing Login": "Login incompleto",
-            "Enter both your RetroAchievements username and password.": "Digite usuário e senha do RetroAchievements.",
-            "Logged In": "Conectado",
-            "Login Failed": "Falha no login",
-            "Not logged in": "Não conectado",
-            "Directory": "Diretório",
-            "Assigned Cards": "Cartões atribuídos",
-            "Slot 1": "Slot 1",
-            "Slot 2": "Slot 2",
-            "Unplugged": "Desconectado",
-            "Slot changes take effect on the next VM boot.": "Mudanças de slot valem no próximo boot da VM.",
-            "Card name": "Nome do cartão",
-            "Create Folder Card": "Criar cartão em pasta",
-            "Create Card": "Criar cartão",
-            "File cards support 8 MB, 16 MB, 32 MB, and 64 MB. Folder cards match the ARMSX2/PCSX2 folder-card behavior and are useful for game-specific saves.": "Cartões de arquivo suportam 8, 16, 32 e 64 MB. Cartões em pasta seguem ARMSX2/PCSX2 e são úteis para saves por jogo.",
-            "No cards found.": "Nenhum cartão encontrado.",
-            "Memory card created.": "Cartão criado.",
-            "Could not create memory card. Check the name, size, or whether it already exists.": "Não foi possível criar o cartão. Verifique nome, tamanho ou se já existe.",
-            "PS2 HDD": "HDD PS2",
-            "Enable DEV9 Virtual HDD": "Ativar HDD virtual DEV9",
-            "Image": "Imagem",
-            "Use Default HDD Image": "Usar imagem HDD padrão",
-            "Matches ARMSX2 Android's DEV9 HDD option. Requires a VM restart.": "Combina com a opção DEV9 HDD do ARMSX2 Android. Requer reiniciar a VM.",
-            "Online / Ethernet": "Online / Ethernet",
-            "Enable DEV9 Ethernet": "Ativar Ethernet DEV9",
-            "Mode": "Modo",
-            "Sockets": "Sockets",
-            "Adapter": "Adaptador",
-            "Intercept DHCP": "Interceptar DHCP",
-            "Log DHCP": "Registrar DHCP",
-            "Log DNS": "Registrar DNS",
-            "Sockets is the iOS-safe DEV9 Ethernet mode exposed here. PCAP bridged/switched modes are compiled out of this iOS build, so they are intentionally not selectable until a real iOS backend exists.": "Sockets é o modo DEV9 seguro no iOS. Modos PCAP ponte/comutado foram removidos desta build até existir um backend real para iOS.",
-            "Tester Notes": "Notas para testers",
-            "Games still need their in-game PS2 network setup. After changing DEV9 settings, use Reset ROM or restart the VM before testing.": "Os jogos ainda precisam da configuração de rede PS2 dentro do jogo. Após mudar DEV9, use Reiniciar jogo ou reinicie a VM antes de testar.",
-            "CPU Recompiler": "Recompilador CPU",
-            "Changes take effect on next VM boot.": "Mudanças valem no próximo boot da VM.",
-            "Boot": "Boot",
-            "Memory": "Memória",
-            "Frame Limiter": "Limitador de FPS",
-            "FPS Target": "FPS alvo",
-            "Speed Target": "Velocidade alvo",
-            "Unlocked": "Sem limite",
-            "NTSC Base Rate": "Taxa base NTSC",
-            "PAL Base Rate": "Taxa base PAL",
-            "FPS Target maps to PCSX2 Normal Speed: 60 FPS is normal NTSC timing, 30 FPS is about 50% speed, and higher values fast-forward. Turning the limiter OFF unlocks speed and can increase heat and battery drain.": "FPS alvo mapeia para velocidade normal do PCSX2: 60 FPS é NTSC normal, 30 FPS é cerca de 50%, e valores maiores aceleram. Desligar o limitador aumenta calor e gasto de bateria.",
-            "Use VU1 Interpreter Preset": "Usar preset interpretador VU1",
-            "Use Full Interpreter Preset": "Usar preset interpretador completo",
-            "GameDB Automatic Fixes": "Correções automáticas GameDB",
-            "Widescreen Patches": "Patches widescreen",
-            "No-Interlacing Patches": "Patches sem entrelaçamento",
-            "GameDB Core Fixes covers timing, clamps, and gamefixes. GameDB Graphics Fixes covers renderer-specific hardware fixes; turn it off globally or per-game if a title looks worse on Metal. PNACH cheats and 60 FPS patches can be imported from the in-game quick menu or from a game's long-press menu.": "GameDB Core cobre timing, clamps e gamefixes. GameDB Graphics cobre fixes do renderer; desligue globalmente ou por jogo se piorar no Metal. PNACH e patches 60 FPS podem ser importados pelo menu rápido ou menu do jogo.",
-            "EE Cycle Rate": "Ciclo EE",
-            "0 = Default. Negative = underclock (stable). Positive = overclock (fast but risky).": "0 = padrão. Negativo = underclock estável. Positivo = overclock rápido mas arriscado.",
-            "Fast CDVD": "CDVD rápido",
-            "Wait Loop Detection": "Detecção de espera",
-            "INTC Stat Hack": "Hack INTC Stat",
-            "VU1 Instant and MTVU are independent now. MTVU can help some games, but keep it off unless a game specifically benefits on iOS.": "VU1 Instant e MTVU agora são independentes. MTVU pode ajudar alguns jogos, mas deixe desligado a menos que o jogo se beneficie no iOS.",
-            "Speedhacks": "Speedhacks",
-            "Reset Emulator to Defaults": "Restaurar emulador",
-            "Usage": "Uso",
-            "App Cache": "Cache do app",
-            "Texture Dumps": "Dumps de textura",
-            "Diagnostic Logs": "Logs de diagnóstico",
-            "State Safety Backups": "Backups de segurança de estado",
-            "Generated Total": "Total gerado",
-            "Refresh Storage Usage": "Atualizar uso de armazenamento",
-            "Cleanup": "Limpeza",
-            "Clear App Cache": "Limpar cache do app",
-            "Clear Texture Dumps": "Limpar dumps de textura",
-            "Clear Diagnostic Logs": "Limpar logs",
-            "Clear All Generated Files": "Limpar arquivos gerados",
-            "This never removes imported games, BIOS files, save states, memory cards, covers, settings, PNACH files, or replacement texture packs.": "Nunca remove jogos importados, BIOS, save states, cartões, capas, ajustes, PNACH ou pacotes de textura.",
-            "Renderer": "Renderizador",
-            "Upscaling": "Upscaling",
-            "Filtering": "Filtragem",
-            "Display": "Tela",
-            "Quality": "Qualidade",
-            "Texture Replacement": "Substituição de textura",
-            "Texture Dumping": "Dump de textura",
-            "VSync": "VSync",
-            "Reset Graphics to Defaults": "Restaurar gráficos",
-            "Load Replacement Textures": "Carregar texturas substitutas",
-            "Dump Replaceable Textures": "Dump de texturas substituíveis",
-            "Dump Mipmaps": "Dump de mipmaps",
-            "Dump During FMV": "Dump durante FMV",
-            "Dump Direct Textures": "Dump de texturas diretas",
-            "Dump Palette Textures": "Dump de texturas de paleta",
-            "Queue Size": "Tamanho da fila",
-            "Higher values reduce frame drops but increase latency.": "Valores maiores reduzem quedas de frames, mas aumentam latência."
-        ]
-    ]
-}
-
 @Observable
 final class SettingsStore: @unchecked Sendable {
     static let shared = SettingsStore()
